Change serialization for graph elements
The previous method would "embed" the objects into eachother-- which meant that update events where a pain. To solve the update problem I simply remove data duplication. This way each struct serializes to a list of the keys it points at-- so the caller can re-assemble the pointers in their own mechanism