Fix bug introduced in 64cc26927b93822d7837f78ade425cde6d76d7bf
When removing the source and dst nodes from the path, it was done after the check that the routes where the same. This mean that routes where always different. In addition-- the first hop in the network is interesting, as it isn't the peer. The goal here is to not have a peer address in the "path"