Support "merging" routes
Since things in the internet don't always return ICMP responses when they drop a packet (rate limiting, etc.)-- its fairly common to get partial traceroutes. Instead of continually replacing the routes-- we can instead attempt to merge them. With this we can (assuming the network is stable) eventually get a complete traceroute by augmenting the one that we have already.

So, for a basic example lets say we had the following route:

1,1|*|3,3,4

and the next response was

1,2,2|*|4,4

Instead of now having a different partial route (1,2,2|*|4,4) we'll end up with a complete route (1,2,3,4).